Extracted Class Responses.java from Response.java
